Fix directory removal logic in empty-dir-remover.py

+++ empty-dir-remover.py
@@ -33,9 +33,12 @@ if path.exists():
                 removed = 0
                 while removed < total:
                     num = int(input("Select directory (number): "))
-                    reference[num].rmdir()
-                    removed += 1
-                    print(f"{reference[num]} removed successfully")
+                    if num in reference.keys():
+                        reference[num].rmdir()
+                        removed += 1
+                        print(f"{reference[num]} removed successfully")
+                    else:
+                        print("Number not in list")
         else:
             print("No empty directories")
     else:
